requirements.txt
python-jose[cryptography]
fastapi
uvicorn[standard]
requests
